+++ src/literalizer/languages/clojure.py
@@ -41,6 +41,7 @@ from literalizer._language import (
     DictFormatConfig,
     FloatSpecialsMixin,
     HeterogeneousBehavior,
+    IdentifierCase,
     LanguageCls,
     OrderedMapFormatConfig,
     PrefixCallStyle,
@@ -294,6 +295,10 @@ class Clojure(metaclass=LanguageCls):
         ERROR = NO_HETEROGENEOUS_BEHAVIOR
 
     heterogeneous_strategies = HeterogeneousStrategies
+
+    identifier_cases: ClassVar[tuple[IdentifierCase, ...]] = (
+        IdentifierCase.KEBAB,
+    )
 
     validate_spec_for_data = no_validate_spec_for_data