+++ tests/integration/cases/call_ref_args_converted_nonsnake/Scala_call.scala
@@ -1,0 +1,15 @@
+object Check {
+def process(data: Any = null, count: Any = null): Any = null
+val myVar = List[Int](
+    1,
+    2,
+    3,
+)
+val myOther = List[Int](
+    4,
+    5,
+    6,
+)
+process(data = myVar, count = 42)
+process(data = myOther, count = 7)
+}