+++ tests/integration/cases/call_ref_args_converted_whole/Rust_call.rs
@@ -1,0 +1,9 @@
+fn main() {
+    fn process<A>(_data: A) {}
+    let my_var = vec![
+        1,
+        2,
+        3,
+    ];
+    process(my_var);
+}